test(e2e): fit the suite on esp32c3's ~172KB JS heap
The heavy test files OOM'd at module load on esp32c3. Split them so each fits, lazy-import the wifi and http module graphs, and skip sections that need more heap than the runtime has via free-heap gates. Heap baselines re-recorded for esp32c3, esp32c5, esp32c6, and the simulator.
