+++ dev/e2e/test/cbor.test.heap-baseline.json
@@ -1,4 +1,7 @@
 {
+  "esp32c3": {
+    "heapDelta": 4050
+  },
   "esp32c5": {
     "heapDelta": 4050
   },
