+++ dev/e2e/test/fs-stream.test.heap-baseline.json
@@ -1,0 +1,14 @@
+{
+  "esp32c3": {
+    "heapDelta": 4612
+  },
+  "esp32c5": {
+    "heapDelta": 4612
+  },
+  "esp32c6": {
+    "heapDelta": 4612
+  },
+  "simulator": {
+    "heapDelta": 4984
+  }
+}
