+++ dev/e2e/test/fs.test.ts
@@ -1,16 +1,4 @@
-import {
-  exists,
-  mkdir,
-  readDir,
-  readFile,
-  readStream,
-  rename,
-  rmdir,
-  stat,
-  unlink,
-  writeFile,
-} from 'mikro/fs'
-import {decodeUtf8, splitLines} from 'mikro/stream'
+import {exists, mkdir, readDir, readFile, rename, rmdir, stat, unlink, writeFile} from 'mikro/fs'
 import {afterEach, assert, describe, test} from 'mikro/test'
 
 // All test paths live under /e2e-fs/ so cleanup is a single recursive walk
@@ -242,58 +230,5 @@ describe('fs: FSError variants', () => {
     const r = writeFile('/app/should-not-write', 'x')
     assert.err(r)
     assert.equal(r.error.name, 'AccessDenied')
-  })
-})
-
-describe('fs: readStream', () => {
-  afterEach(cleanup)
-
-  test('yields chunks until EOF', async () => {
-    setup()
-    const payload = 'The quick brown fox jumps over the lazy dog.'
-    assert.ok(writeFile(`${ROOT}/fox.txt`, payload))
-
-    const r = readStream(`${ROOT}/fox.txt`, {chunkSize: 8})
-    assert.ok(r)
-    const decoder = new TextDecoder()
-    let out = ''
-    for await (const chunk of r.value) {
-      assert.ok(chunk)
-      out += decoder.decode(chunk.value)
-    }
-    assert.equal(out, payload)
-  })
-
-  test('composes with decodeUtf8 and splitLines', async () => {
-    setup()
-    assert.ok(writeFile(`${ROOT}/lines.txt`, 'one\ntwo\nthree\n'))
-
-    const r = readStream(`${ROOT}/lines.txt`)
-    assert.ok(r)
-    const lines: string[] = []
-    for await (const line of splitLines(decodeUtf8(r.value))) {
-      assert.ok(line)
-      lines.push(line.value)
-    }
-    assert.deepEqual(lines, ['one', 'two', 'three'])
-  })
-
-  test('empty file yields no chunks', async () => {
-    setup()
-    assert.ok(writeFile(`${ROOT}/empty.txt`, ''))
-
-    const r = readStream(`${ROOT}/empty.txt`)
-    assert.ok(r)
-    let chunks = 0
-    for await (const _ of r.value) {
-      chunks++
-    }
-    assert.equal(chunks, 0)
-  })
-
-  test('missing file returns FSError.NotFound', () => {
-    const r = readStream(`${ROOT}/nope.txt`)
-    assert.err(r)
-    assert.equal(r.error.name, 'NotFound')
   })
 })
