+++ dev/e2e/test/observable.test.heap-baseline.json
@@ -1,4 +1,7 @@
 {
+  "esp32c3": {
+    "heapDelta": 4473
+  },
   "esp32c5": {
     "heapDelta": 4473
   },
