+++ dev/e2e/test/udp.test.heap-baseline.json
@@ -1,11 +1,14 @@
 {
+  "esp32c3": {
+    "heapDelta": 4073
+  },
   "esp32c5": {
-    "heapDelta": -352
+    "heapDelta": 144
   },
   "esp32c6": {
-    "heapDelta": -352
+    "heapDelta": 144
   },
   "simulator": {
-    "heapDelta": 568
+    "heapDelta": 4313
   }
 }
